# billingwebapp/docs/openapi-ai-v1.yaml
openapi: 3.0.3
info:
  title: Clinic AI Integration API
  version: 1.0.0
  description: Secure server-to-server API for the separate Call Assistant. Direct database access is unsupported.
servers:
  - url: /api/v1/ai
security:
  - bearerAuth: []
paths:
  /auth/token:
    post:
      summary: Exchange AI client credentials for a short-lived opaque token
      security: [{basicAuth: []}, {}]
      requestBody:
        content:
          application/json: {schema: {$ref: '#/components/schemas/TokenRequest'}}
          application/x-www-form-urlencoded: {schema: {$ref: '#/components/schemas/TokenRequest'}}
      responses: {'200': {$ref: '#/components/responses/Success'}, '401': {$ref: '#/components/responses/Unauthorized'}, '429': {$ref: '#/components/responses/RateLimited'}}
  /integration/status:
    get:
      summary: Validate connectivity and authenticated client context
      x-required-scope: clinic:read
      responses: {'200': {$ref: '#/components/responses/Success'}, '403': {$ref: '#/components/responses/Forbidden'}}
...
  /documents/{token}:
    get:
      summary: Consume a short-lived one-time document token
      security: []
      parameters: [{name: token, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Private report stream}, '403': {$ref: '#/components/responses/Forbidden'}}
...
components:
  securitySchemes:
    basicAuth: {type: http, scheme: basic}
    bearerAuth: {type: http, scheme: bearer, bearerFormat: opaque}
...
  schemas:
    TokenRequest:
      type: object
      required: [client_id, client_secret]
      properties: {client_id: {type: string}, client_secret: {type: string, format: password}, scope: {type: string}}
...
    Success:
      type: object
      required: [success, data, request_id]
      properties: {success: {type: boolean, enum: [true]}, data: {type: object}, request_id: {type: string}}
    Error:
      type: object
      required: [success, error, request_id]
      properties:
        success: {type: boolean, enum: [false]}
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
        request_id: {type: string}
  responses:
    Success: {description: Standard success envelope, content: {application/json: {schema: {$ref: '#/components/schemas/Success'}}}}
...
    Unauthorized: {description: Invalid or expired authentication, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
    Forbidden: {description: Scope or ownership check failed, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    RateLimited: {description: Rate limit exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}